Fix impulse shortcuts, lowercase i and inline code
Lowercase the "i" in I+DUP. Because uppercase I (i) looks the same as lowercase l (L) (at least on the web page). The gif animations show the shortcut as"i + DUP" and the last section "Creating an Impulse Chord on the CharaChorder X" shows the shortcut as "i+ESC".