Remove embedded action codes
The embedded Google sheet, with the action codes, causes some issues:

On the Device Manager page:
- it scrolls to about the vertical middle when it loads
- it takes about 10 seconds to load instead of 1 second without the embed

The SerialAPI page also has the action codes embedded, but it's embedded as a preview. The Device Manager embedded it in edit mode (it shows the sheets menus and toolbars).

Changing the Device Manager page to an embedded preview, stops the auto scrolling, but sometimes both the Device Manager page and SerialAPI page takes a couple of seconds to load.

Another issue is that the embedded Google sheet has a fixed width, and it requires horizontal scrolling to view the last columns.

Therefore, the embedding was removed from both the Device Manager and SerialAPI pages, leaving a link to open the action codes in a full page.